looks like working

+++ CmdParameters.cpp
@@ -16,7 +16,7 @@ void CmdParameters::init(int argc, char* argv[]) {
 };
 
 bool CmdParameters::has(std::string strValue) {
-	for (int n = 0; n < vParameters.size(); n++) {
+	for (int n = 0; n < (int)vParameters.size(); n++) {
 		if (vParameters[n].compare(strValue) == 0) return true;
 	};
 	return false;
@@ -24,8 +24,8 @@ bool CmdParameters::has(std::string strValue) {
 
 
 std::string CmdParameters::getArg(std::string strValue) {
-	for (int n = 0; n < vParameters.size(); n++) {
-		if ((vParameters[n].compare(strValue) == 0) && (n < (vParameters.size() - 1))) return vParameters[n + 1];
+	for (int n = 0; n < (int)vParameters.size(); n++) {
+		if ((vParameters[n].compare(strValue) == 0) && (n < ((int)vParameters.size() - 1))) return vParameters[n + 1];
 	};
 	return emptyString;
 };
